Add a temporary getAllObjects() in NContinuous2D | make discretizations double

+++ mason/sim/app/dflockers3/DFlockers.java
@@ -40,7 +40,7 @@ public class DFlockers extends SimState {
 		{
 			int[] aoi = new int[] { (int) neighborhood, (int) neighborhood };
 			int[] size = new int[] { (int) width, (int) height };
-			int[] discretizations = new int[] { (int) (neighborhood / 1.5), (int) (neighborhood / 1.5) };
+			double[] discretizations = new double[] { neighborhood / 1.5, neighborhood / 1.5 };
 
 			partition = DNonUniformPartition.getPartitionScheme(size, true, aoi);
 			partition.initUniformly(null);
